Updates & fixes to books views & controllers

+++ app/Http/Controllers/BooksController.php
@@ -10,6 +10,11 @@ use Illuminate\Support\Facades\Storage;
 class BooksController extends Controller
 {
     /**
+     * ISBN regex to check valid isbn-13 and isbn-10 numbers
+     */
+    private $isbnRegex = "^(?:(?:978-?)?\d{1,5}-?\d{1,7}-?\d{1,9}-?\d|(?:(?:978)?\d{9}[0-9X]))$";
+
+    /**
      * Display a listing of the resource.
      */
     public function index()
@@ -36,7 +41,7 @@ class BooksController extends Controller
     /**
      * Display all the resources for multi selection
      */
-    public function multi_select()
+    public function multiSelect()
     {
         $books = Book::paginate(3);
 
@@ -69,12 +74,12 @@ class BooksController extends Controller
     public function store(Request $request)
     {
         $validated = $request->validate([
-            "name" => "required|string",
-            "isbn" => "required|string",
-            "desc" => "nullable|string",
+            "name" => "required|string|max:500",
+            "isbn" => ["required", "string", "regex:/$this->isbnRegex/"],
+            "desc" => "nullable|string|max:1000",
             "price" => "required|numeric",
             "stock" => "required|numeric",
-            "rate" => "required|numeric",
+            "rate" => "required|numeric|max:10",
             "img" => "nullable|image|max:10240"
         ]);
 
@@ -101,14 +106,15 @@ class BooksController extends Controller
      */
     public function show(int $id)
     {
-        $book = Book::query()->where("id", "=", $id)->first();
+        $book = Book::findOrFail($id);
 
-        if (!$book) {
-            abort(404, "Resource does not exist!");
-        }
+        $genres = $book->genres()->get();
+        $authors = $book->authors()->get();
 
         return view("books.show")->with([
-            "book" => $book
+            "book" => $book,
+            "genres" => $genres,
+            "authors" => $authors
         ]);
     }
 
@@ -117,11 +123,7 @@ class BooksController extends Controller
      */
     public function edit(int $id)
     {
-        $book = Book::query()->where("id", "=", $id)->first();
-
-        if (!$book) {
-            abort(404, "Resource does not exist!");
-        }
+        $book = Book::findOrFail($id);
 
         return view("books.edit")->with([
             "book" => $book
@@ -133,19 +135,15 @@ class BooksController extends Controller
      */
     public function update(Request $request, int $id)
     {
-        $book = Book::query()->where("id", "=", $id)->first();
-
-        if (!$book) {
-            abort(404, "Resource does not exist!");
-        }
+        $book = Book::findOrFail($id);
 
         $validated = $request->validate([
-            "isbn" => "nullable|string",
-            "name" => "nullable|string",
-            "desc" => "nullable|string",
+            "isbn" => ["nullable", "string", "regex:/$this->isbnRegex/"],
+            "name" => "nullable|string|max:500",
+            "desc" => "nullable|string|max:1000",
             "price" => "nullable|numeric",
             "stock" => "nullable|numeric",
-            "rate" => "nullable|numeric",
+            "rate" => "nullable|numeric|max:10",
             "img" => "nullable|image|max:10240"
         ]);
 
@@ -182,10 +180,14 @@ class BooksController extends Controller
         $book = Book::findOrFail($id);
 
         $book->ratings()->delete();
+        $book->genres()->detach();
         $book->authors()->detach();
         $book->items()->delete();
 
-        Storage::disk('public')->delete($book->img);
+        if ($book->img) {
+            Storage::disk('public')->delete($book->img);
+        }
+            
         $book->delete();
 
         return response(null);
